support multiple polygon tables
a record now has an array of polygon ids (one per polygon table) which are populated
by the netacq plugin. the order of the polygon ids corresponds to the order that
the polygon tables were loaded.  ipmeta_provider_netacq_edge_get_polygon_tables
can be used to retrieve an array (with the same ordering) of polygon tables
and corresponding polygon lookup information.

given an array of polygon tables, `poly_tables`, and a record `rec`, the fqid
for the `i`th table can be retrieved with something like:
```
poly_tables[i]->polygons[rec->polygon_ids[i]]->fqid;
```