First pass at a re-structure to share a single DS for all providers
The idea is that, in situations where multiple providers are being
used, it is more efficient to do one lookup and return N records
than it is to do N lookups for 1 record each.

This obviously has plenty of ramifications:
 * API changes for several functions that are not backwards
   compatible.
 * Interval Tree data structure doesn't really work with this
   approach (at least not without *more* API changes) since we
   can't "update" an existing node.
 * Added new concepts such as provider masks which limit lookups
   to a specific set of providers.
 * Apps will need to handle config of the data structure, rather
   than it being a per-provider option.
 * Providers cannot mix and match data structures, although I
   doubt this was common to begin with.

Only the maxmind provider has been updated for the new API so far,
but will do the others once I've done some basic testing.

Both bigarray and patricia data structures *should* work with
the new API, but require more testing to confirm.