Still allow -D option to be given to each provider, but ignore it.
This is to make life a bit easier when moving older deployments to
use the new API -- without this change, any existing configuration
would be invalid and cause a fatal error.

Instead, we now print a warning about how the option is no longer
valid and carry on, so users can fix their config at their leisure.