Remove default provider feature
This was a half-baked, never-used idea that was designed for a case
where a libipmeta instance was managed centrally and users could ask for
a "default" provider if they didn't have a preference. This only makes
sense when there are multiple providers that perform the same function
(e.g., geolocation), and doesn't help at all if there is e.g., a
geolocation provider and a pfx2as provider enabled. Since this was never
used, it just adds unncessary complexity to the interface.