Fix returning multiple results from patricia lookup
Since the v2 redesign to share a single DS across all providers, a
prefix lookup in the patricia DS that should have returned multiple
results from a provider returned only one, because as soon as it found
one it set a flag meaning "already found a result for this provider" and
stopped looking for any other results for that provider.

Also restore the result order to preorder as in v1.